reconfigured the src directory to work properly for builds. Moved tests to hopefully work with this layout
