+++ src/privateai_client/components/pai_uris.py
