+++ src/privateai_client/tests/test_request_objects.py
@@ -1,5 +1,7 @@
 import pytest
-from src.components import *
+
+from ..components import *
+
 
 # File Tests
 def test_file_initializer():
